Added HasOption trait Added trait to Select and MultiSelect
